rustlings_tutorials/requirements.txt
pillow
requests
cairosvg
